添加效果用例pagination-type
